Add server URL to the cmd config parameters
